# faster sparse(::Diagonal)