Missing method `sparse(::Diagonal)`:

| Size          | master                                  | PR                                    | speedup |
|---------------|-----------------------------------------|---------------------------------------|---------|
| 100x100       | 29.076 μs (7 allocations: 4.45 KiB)     | 210.451 ns (5 allocations: 1.86 KiB)  | 140     |
| 1000x1000     | 2.600 ms (7 allocations: 39.89 KiB)     | 1.450 μs (5 allocations: 16.05 KiB)   | 2k    |
| 10000x10000   | 283.301 ms (12 allocations: 391.22 KiB) | 14.472 μs (7 allocations: 156.58 KiB) | 20k     |
| 100000x100000 | 27.050 s (12 allocations: 3.82 MiB)     | 169.198 μs (7 allocations: 1.53 MiB)  | 160k    |