@fredrikekre 2017-08-31
https://github.com/JuliaLang/julia/blob/b729e58dff9c9cfa2640b8c4dcc71dbc91b29871/test/sparse/sparse.jl#L1254-L1255 and also indirectly on some other places. But might as well add tests to this block https://github.com/JuliaLang/julia/blob/b729e58dff9c9cfa2640b8c4dcc71dbc91b29871/test/sparse/sparse.jl#L1295-L1308