@fredrikekre 2017-09-01
Updated, and updated the benchmarks in top post too.